+++ yoga/Utils.h
@@ -38,19 +38,19 @@
 
 struct YGCollectFlexItemsRowValues {
   uint32_t itemsOnLine;
-  float sizeConsumedOnCurrentLine;
-  float totalFlexGrowFactors;
-  float totalFlexShrinkScaledFactors;
+  double sizeConsumedOnCurrentLine;
+  double totalFlexGrowFactors;
+  double totalFlexShrinkScaledFactors;
   uint32_t endOfLineIndex;
   std::vector<YGNodeRef> relativeChildren;
-  float remainingFreeSpace;
+  double remainingFreeSpace;
   // The size of the mainDim for the row after considering size, padding, margin
   // and border of flex items. This is used to calculate maxLineDim after going
   // through all the rows to decide on the main axis size of owner.
-  float mainDim;
+  double mainDim;
   // The size of the crossDim for the row after considering size, padding,
   // margin and border of flex items. Used for calculating containers crossSize.
-  float crossDim;
+  double crossDim;
 };
 
 bool YGValueEqual(const YGValue& a, const YGValue& b);
@@ -60,27 +60,27 @@ inline bool YGValueEqual(
   return YGValueEqual((YGValue) a, (YGValue) b);
 }
 
-// This custom float equality function returns true if either absolute
+// This custom double equality function returns true if either absolute
 // difference between two floats is less than 0.0001f or both are undefined.
-bool YGFloatsEqual(const float a, const float b);
+bool YGFloatsEqual(const double a, const double b);
 
 bool YGDoubleEqual(const double a, const double b);
 
-float YGFloatMax(const float a, const float b);
+double YGFloatMax(const double a, const double b);
 
 YGFloatOptional YGFloatOptionalMax(
     const YGFloatOptional op1,
     const YGFloatOptional op2);
 
-float YGFloatMin(const float a, const float b);
+double YGFloatMin(const double a, const double b);
 
-// This custom float comparison function compares the array of float with
-// YGFloatsEqual, as the default float comparison operator will not work(Look
+// This custom double comparison function compares the array of double with
+// YGFloatsEqual, as the default double comparison operator will not work(Look
 // at the comments of YGFloatsEqual function).
 template <std::size_t size>
 bool YGFloatArrayEqual(
-    const std::array<float, size>& val1,
-    const std::array<float, size>& val2) {
+    const std::array<double, size>& val1,
+    const std::array<double, size>& val2) {
   bool areEqual = true;
   for (std::size_t i = 0; i < size && areEqual; ++i) {
     areEqual = YGFloatsEqual(val1[i], val2[i]);
@@ -89,7 +89,7 @@ bool YGFloatArrayEqual(
 }
 
 // This function returns 0 if YGFloatIsUndefined(val) is true and val otherwise
-float YGFloatSanitize(const float val);
+double YGFloatSanitize(const double val);
 
 YGFlexDirection YGFlexDirectionCross(
     const YGFlexDirection flexDirection,
@@ -102,12 +102,12 @@ inline bool YGFlexDirectionIsRow(const YGFlexDirection flexDirection) {
 
 inline YGFloatOptional YGResolveValue(
     const YGValue value,
-    const float ownerSize) {
+    const double ownerSize) {
   switch (value.unit) {
     case YGUnitPoint:
       return YGFloatOptional{value.value};
     case YGUnitPercent:
-      return YGFloatOptional{value.value * ownerSize * 0.01f};
+      return YGFloatOptional{value.value * ownerSize * 0.01};
     default:
       return YGFloatOptional{};
   }
@@ -115,7 +115,7 @@ inline YGFloatOptional YGResolveValue(
 
 inline YGFloatOptional YGResolveValue(
     yoga::detail::CompactValue value,
-    float ownerSize) {
+    double ownerSize) {
   return YGResolveValue((YGValue) value, ownerSize);
 }
 
@@ -140,7 +140,7 @@ inline YGFlexDirection YGResolveFlexDirection(
 
 inline YGFloatOptional YGResolveValueMargin(
     yoga::detail::CompactValue value,
-    const float ownerSize) {
+    const double ownerSize) {
   return value.isAuto() ? YGFloatOptional{0} : YGResolveValue(value, ownerSize);
 }
